improve ADC noise rejection and add self-calibration
ADC filter pipeline: add sliding median(5) for spike rejection, increase
oversampling from 50→200 samples (10s window, √200≈14× noise reduction),
lower EMA alpha 0.8→0.1. Promote raw_adc to float throughout to preserve
sub-integer precision after filtering. Lower volume report threshold 3→1 L.

Self-calibration module collects warming-phase ADC/temperature pairs via
online linear regression and exposes the calculated slope via RPC and web UI.
Fix missing mgos_bme280.h include in sensor_calibration.c. Fix pressure
average filter struct type (harmonic→arithmetic).

README updated to document the three-stage filter pipeline and self-calibration
readiness criteria.

+++ src/main.c
@@ -102,7 +102,7 @@ struct sensor_info
 struct sensor_raw
 {
   time_t    timestamp;
-  uint16_t  tank_pressure_adc;
+  float     tank_pressure_adc;
   uint16_t  counter_count;
   float     counter_frequency;
 } sensor_raw = {
@@ -157,7 +157,7 @@ const struct mbuf *getRawAsJSON(struct mbuf *buffer)
   json_printf(  &json_result,
                 "{"
                 "timestamp: %d,"
-                "tank_pressure_adc: %d,"
+                "tank_pressure_adc: %4.1f,"
                 "tank_overflow_count: %d,"
                 "tank_overflow_frequency: %3.1f"
                 "}",
